Adding thoughts on security in giles 2
I think I have most of the scheme planned out now. Symmetric encryption
is going to be at the key of this design, so I will need to investigate
and benchmark some AES (or others?) implementations for go. With any
luck I won't have to look outside of the standard library, though
linking in a good C implementation would be a nice exercise.