Outline EphemeralKey methods in PermissionsManager
Currently, all of PermissionsManager is behind an interface, but there
are some features such as the ephemeral key cache that we want to
preserve independently of the db provider for the permissions manager
for performance and architectural reasons. This suggests that we should
move to a model where the permissions manager is an actual struct, and
the db provider will implement lower level functions.