Fix SmapProperties pointer reference and mongo key
We switched to using &SmapProperties inside the SmapMessage struct to
better detect nil properties when marshalling the data structure.
There is probably a defined way to have a custom MarshalJSON interface
implementer to return an encoding that registers as 'omitempty' on an
encapsulating struct, but I can't find it.

Due to how everything is stored inside mongo, we have to pay a little
more attention on the formatting of keys as they come out of the
metadataStore interface implementation. Need to write up some comments
on that