Starting on republish/subscriber implementation
Trying to go for a slightly more predictable model this time. Going to
use channels (buffered suggested, but optional) as the delivery
mechanism in order to guarantee order. Also using this opportunity to
fix up the implementation of republish so that it delivers the full
query results upon initial subscription, and offers incremental updates
where it can. This will definitely be much more involved than last time,
but it should go fairly quickly