Add ValidTimestamp method for timeseries databases
One of the big problems we've had is timestamps being out of the valid
range for btrdb/quasar and thus crashing it after enough of them have
been written. It would be nice if we could stop those timestamps from
even getting to the database, which I'm doing with the help of
github.com/SoftwareDefinedBuildings/btrdb/qtree.MaximumTime. Still
remains to be seen where this will get used in the code