Do cache invalidation on changing metadata
Specifically, the uot/uom caches get invalidated when new metadata comes
in for a uuid. The invalidation happens on a uuid by uuid basis. Moving
forward, we could probably make the invalidation a little smarter, e.g.
by only invalidating if UnitofTime or UnitofMeasure are mentioned or
only invalidating if they have changed (in which case we could just use
the in-place edit on the ccache)